refactor(controller): encapsulate it into a Stream
With this Lambda form of thing, WriterStream and
Controller need the same Command enum. I dont
know if this is good or not. But it feels better
than just handing over a Sender<Command>

Note the lambda needs to also have the + use<'b>
because of the default elisian &self 's lifetime
is just applied to all outputs. but i cloned it
so it should be necessary.

also the impl is because im returning an impl
of a trait. i cant use generic since that would
rely on it being hte caller controlling it. i
think idk